feat: extracted setupMatchRoles to a single function
